inprovements in utils
