+++ src/utils/__init__.py
@@ -1,3 +1,2 @@
-from .dominant_color import dominant_color
-from .assets import Colors, Emojis, MediaUrl
-from .newassets import *
+from .colors import *
+from .emojis import Emojis
