+++ src/utils/emojis.py
@@ -1,0 +1,32 @@
+class Emojis:
+	#status
+	ERROR = '<:X_:1067469281475965009>'
+	NEUTRAL = '<:neutral:1067469339885830174>'
+	SUCCESS = '<:check:1067469415706263562>'
+
+	#geral
+	DOT = '<:dot:1067471950491955210>'
+	TRASH = '<:trash:1066030051411361873>'
+	BOT_TAG = '<:bot_tag:1067461733150621786>'
+
+	#social
+	GITHUB = '<:github:1067459903557468200>'
+
+	#tools
+	PYTHON = '<:python:1017975869777645639>'
+	DISNAKE = '<:disnake:1017292954278309888>'
+
+	#UNO
+	BLOCK = '<:block:1066046626671628418>'
+	REVERSE = '<:reverse:1066045215795843212>'
+
+	#foda's
+	MONOCULO = '<:monoculo:1066057929247179005>'
+
+	#animated
+	BOOST = '<a:boost:1066061313874329740>'
+	DANCE = '<a:dance:1066043138491306134>'
+
+	#dual
+	NEW = '<:new1:1067472996576542771><:new2:1067473040100839464>'
+	BETA = '<:beta1:1017294298275926027><:beta2:1017294345612824618>'
